캡스톤/window-app/server/requirements.txt
fastapi
uvicorn[standard]
pydantic
paho-mqtt
requests
python-dotenv
firebase-admin
